zcash_client_sqlite: Move transparent-inputs wallet methods into the `wallet::transparent` module.
